+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 #### 2017.08.31
 1. Now axes handle can be passed to colormesh2d()
-2. Added replist to mpes.utils
+2. Added replist, shuffleaxis to mpes.utils
 3. Added fit_parameter_plot() to mpes.visualization (built on top of colormesh2d)
 
 #### 2017.08.29